fix(server): breakpoints can be used by MediaObserver
